requirements.txt
python-dotenv
pandas
numpy
requests
beautifulsoup4
jupyter
tqdm
openai
